tip_calculator
In this project, I had to develop a tip calculator. I did that by getting some inputs from the user such as the total bill, the tip they want to give, and how many people to split the bill with. After getting the necessary inputs, the output will be the price of the bill for each person with a tip included.